Convert PlayerVideoActivity into PreviewVideoActivity
