+++ src/com/owncloud/android/ui/activity/FileDisplayActivity.java
@@ -80,7 +80,6 @@ import com.owncloud.android.ui.fragment.FileFragment;
 import com.owncloud.android.ui.fragment.OCFileListFragment;
 import com.owncloud.android.ui.fragment.TaskRetainerFragment;
 import com.owncloud.android.ui.helpers.UriUploader;
-import com.owncloud.android.ui.preview.PlayerVideoActivity;
 import com.owncloud.android.ui.preview.PreviewAudioFragment;
 import com.owncloud.android.ui.preview.PreviewImageActivity;
 import com.owncloud.android.ui.preview.PreviewImageFragment;
@@ -364,9 +363,9 @@ public class FileDisplayActivity extends HookActivity
                     file.getLastSyncDateForProperties() > 0  // temporal fix
                     ) {
                 int startPlaybackPosition =
-                    getIntent().getIntExtra(PlayerVideoActivity.EXTRA_START_POSITION, 0);
+                    getIntent().getIntExtra(PreviewVideoActivity.EXTRA_START_POSITION, 0);
                 boolean autoplay =
-                    getIntent().getBooleanExtra(PlayerVideoActivity.EXTRA_AUTOPLAY, true);
+                    getIntent().getBooleanExtra(PreviewVideoActivity.EXTRA_AUTOPLAY, true);
 
                 if (PreviewAudioFragment.canBePreviewed(file)) {
                     secondFragment = PreviewAudioFragment.newInstance(
