--- src/com/owncloud/android/ui/preview/PlayerVideoActivity.java
@@ -1,413 +1,0 @@
-/**
- * ownCloud Android client application
- *
- * @author David A. Velasco
- * Copyright (C) 2016 ownCloud GmbH.
- * <p>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- * <p>
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * <p>
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.owncloud.android.ui.preview;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Handler;
-import android.util.Base64;
-import android.view.KeyEvent;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.view.Window;
-import android.view.WindowManager;
-import android.widget.ImageButton;
-import android.widget.ProgressBar;
-import android.widget.Toast;
-
-import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.DefaultLoadControl;
-import com.google.android.exoplayer2.ExoPlaybackException;
-import com.google.android.exoplayer2.ExoPlayer;
-import com.google.android.exoplayer2.ExoPlayerFactory;
-import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.Timeline;
-import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
-import com.google.android.exoplayer2.mediacodec.MediaCodecRenderer.DecoderInitializationException;
-import com.google.android.exoplayer2.mediacodec.MediaCodecUtil.DecoderQueryException;
-import com.google.android.exoplayer2.source.BehindLiveWindowException;
-import com.google.android.exoplayer2.source.ExtractorMediaSource;
-import com.google.android.exoplayer2.source.MediaSource;
-import com.google.android.exoplayer2.source.TrackGroupArray;
-import com.google.android.exoplayer2.trackselection.AdaptiveVideoTrackSelection;
-import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
-import com.google.android.exoplayer2.trackselection.MappingTrackSelector.MappedTrackInfo;
-import com.google.android.exoplayer2.trackselection.TrackSelection;
-import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
-import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
-import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
-import com.google.android.exoplayer2.upstream.HttpDataSource;
-import com.google.android.exoplayer2.util.Util;
-import com.owncloud.android.MainApp;
-import com.owncloud.android.R;
-import com.owncloud.android.lib.common.OwnCloudAccount;
-import com.owncloud.android.lib.common.OwnCloudBasicCredentials;
-import com.owncloud.android.lib.common.OwnCloudCredentials;
-import com.owncloud.android.lib.common.OwnCloudSamlSsoCredentials;
-import com.owncloud.android.lib.common.accounts.AccountUtils;
-import com.owncloud.android.lib.common.utils.Log_OC;
-import com.owncloud.android.ui.activity.FileActivity;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.ExecutionException;
-
-/**
- * An activity that plays media using {@link SimpleExoPlayer}.
- */
-public class PlayerVideoActivity extends FileActivity implements   ExoPlayer.EventListener {
-
-    private static final DefaultBandwidthMeter BANDWIDTH_METER = new DefaultBandwidthMeter();
-
-    private Handler mainHandler;
-    private SimpleExoPlayerView simpleExoPlayerView;
-
-    private SimpleExoPlayer player;
-    private DefaultTrackSelector trackSelector;
-    private boolean playerNeedsSource;
-
-
-    /** Key to receive a flag signaling if the video should be started immediately */
-    public static final String EXTRA_AUTOPLAY = "AUTOPLAY";
-
-    /** Key to receive the position of the playback where the video should be put at start */
-    public static final String EXTRA_START_POSITION = "START_POSITION";
-
-    private boolean mAutoplay; // when 'true', the playback starts immediately with the activity
-    private long mPlaybackPosition; // continue the playback in the specified position
-
-    // Activity lifecycle
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        // Enable full screen
-        this.requestWindowFeature(Window.FEATURE_NO_TITLE); //Remove title bar
-        this.getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-                WindowManager.LayoutParams.FLAG_FULLSCREEN); //Remove notification bar
-
-        clearResumePosition();
-
-        setContentView(R.layout.video_preview);
-
-        simpleExoPlayerView = (SimpleExoPlayerView) findViewById(R.id.video_player);
-
-        // Hide sync bar
-        ProgressBar syncProgressBar = (ProgressBar) findViewById(R.id.syncProgressBar);
-        syncProgressBar.setVisibility(View.GONE);
-
-        // Hide full screen button
-        ImageButton fullScreen = (ImageButton) findViewById(R.id.fullscreen_button);
-        fullScreen.setVisibility(View.GONE);
-
-        // Show exit full screen button
-        ImageButton exitFullScreen = (ImageButton) findViewById(R.id.exit_fullscreen_button);
-        exitFullScreen.setVisibility(View.VISIBLE);
-
-        exitFullScreen.setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                releasePlayer();
-                finish();
-            }
-        });
-
-
-        Bundle extras = getIntent().getExtras();
-
-        mAutoplay = extras.getBoolean(EXTRA_AUTOPLAY);
-        mPlaybackPosition = extras.getLong(EXTRA_START_POSITION);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions,
-                                           int[] grantResults) {
-        if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-            preparePlayer();
-        } else {
-//          // permission denied --> do nothing
-            finish();
-        }
-    }
-
-    // Activity input
-
-    @Override
-    public boolean dispatchKeyEvent(KeyEvent event) {
-        // Show the controls on any key event.
-        simpleExoPlayerView.showController();
-        // If the event was not handled then see if the player view can handle it as a media key event.
-        return super.dispatchKeyEvent(event) || simpleExoPlayerView.dispatchMediaKeyEvent(event);
-    }
-
-
-    private void preparePlayer() {
-
-        // Create a default TrackSelector
-        mainHandler = new Handler();
-        TrackSelection.Factory videoTrackSelectionFactory =
-                new AdaptiveVideoTrackSelection.Factory(BANDWIDTH_METER);
-        trackSelector = new DefaultTrackSelector(videoTrackSelectionFactory);
-        player = ExoPlayerFactory.newSimpleInstance(this, trackSelector, new DefaultLoadControl());
-        player.addListener(this);
-        simpleExoPlayerView.setPlayer(player);
-        player.seekTo(mPlaybackPosition);
-        player.setPlayWhenReady(mAutoplay);
-
-        try {
-
-            // If the file is already downloaded, reproduce it locally
-            Uri uri = getFile().isDown() ? getFile().getStorageUri() :
-                    Uri.parse(AccountUtils.constructFullURLForAccount(this, getAccount()) +
-                            Uri.encode(getFile().getRemotePath(), "/"));
-
-            DataSource.Factory mediaDataSourceFactory = buildDataSourceFactory(true);
-
-            MediaSource mediaSource = buildMediaSource(mediaDataSourceFactory, uri);
-
-            player.prepare(mediaSource);
-
-        } catch (AccountUtils.AccountNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private MediaSource buildMediaSource(DataSource.Factory mediaDataSourceFactory, Uri uri) {
-        return new ExtractorMediaSource(uri, mediaDataSourceFactory, new DefaultExtractorsFactory(),
-                mainHandler, null);
-    }
-
-    private void releasePlayer() {
-        if (player != null) {
-            mAutoplay = player.getPlayWhenReady();
-            updateResumePosition();
-            player.release();
-            player = null;
-            trackSelector = null;
-        }
-    }
-
-    private void updateResumePosition() {
-        mPlaybackPosition = player.isCurrentWindowSeekable() ? Math.max(0, player.getCurrentPosition())
-                : C.TIME_UNSET;
-    }
-
-    private void clearResumePosition() {
-        mPlaybackPosition = C.TIME_UNSET;
-    }
-
-    // ExoPlayer.EventListener implementation
-
-    @Override
-    public void onLoadingChanged(boolean isLoading) {
-        // Do nothing.
-    }
-
-    @Override
-    public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
-        // Do nothing.
-    }
-
-    @Override
-    public void onPositionDiscontinuity() {
-        if (playerNeedsSource) {
-            // This will only occur if the user has performed a seek whilst in the error state. Update the
-            // resume position so that if the user then retries, playback will resume from the position to
-            // which they seeked.
-            updateResumePosition();
-        }
-    }
-
-    @Override
-    public void onTimelineChanged(Timeline timeline, Object manifest) {
-        // Do nothing.
-    }
-
-    @Override
-    public void onPlayerError(ExoPlaybackException e) {
-        String errorString = null;
-        if (e.type == ExoPlaybackException.TYPE_RENDERER) {
-            Exception cause = e.getRendererException();
-            if (cause instanceof DecoderInitializationException) {
-                // Special case for decoder initialization failures.
-                DecoderInitializationException decoderInitializationException =
-                        (DecoderInitializationException) cause;
-                if (decoderInitializationException.decoderName == null) {
-                    if (decoderInitializationException.getCause() instanceof DecoderQueryException) {
-//                        errorString = getString(R.string.error_querying_decoders);
-                    } else if (decoderInitializationException.secureDecoderRequired) {
-//                        errorString = getString(R.string.error_no_secure_decoder,
-//                                decoderInitializationException.mimeType);
-                    } else {
-//                        errorString = getString(R.string.error_no_decoder,
-//                                decoderInitializationException.mimeType);
-                    }
-                } else {
-//                    errorString = getString(R.string.error_instantiating_decoder,
-//                            decoderInitializationException.decoderName);
-                }
-            }
-        }
-        if (errorString != null) {
-            showToast(errorString);
-        }
-        playerNeedsSource = true;
-        if (isBehindLiveWindow(e)) {
-            clearResumePosition();
-            preparePlayer();
-        } else {
-            updateResumePosition();
-        }
-    }
-
-    @Override
-    public void onTracksChanged(TrackGroupArray trackGroups, TrackSelectionArray trackSelections) {
-        // Do nothing
-    }
-
-    private void showToast(String message) {
-        Toast.makeText(getApplicationContext(), message, Toast.LENGTH_LONG).show();
-    }
-
-    private static boolean isBehindLiveWindow(ExoPlaybackException e) {
-        if (e.type != ExoPlaybackException.TYPE_SOURCE) {
-            return false;
-        }
-        Throwable cause = e.getSourceException();
-        while (cause != null) {
-            if (cause instanceof BehindLiveWindowException) {
-                return true;
-            }
-            cause = cause.getCause();
-        }
-        return false;
-    }
-
-    /**
-     * Returns a new DataSource factory.
-     *
-     * @param useBandwidthMeter Whether to set {@link #BANDWIDTH_METER} as a listener to the new
-     *     DataSource factory.
-     * @return A new DataSource factory.
-     */
-    private DataSource.Factory buildDataSourceFactory(boolean useBandwidthMeter) {
-        return buildDataSourceFactory(useBandwidthMeter ? BANDWIDTH_METER : null);
-    }
-
-    private DataSource.Factory buildDataSourceFactory(DefaultBandwidthMeter bandwidthMeter) {
-        return new DefaultDataSourceFactory(this, bandwidthMeter,
-                buildHttpDataSourceFactory(bandwidthMeter));
-    }
-
-    private HttpDataSource.Factory buildHttpDataSourceFactory(DefaultBandwidthMeter bandwidthMeter) {
-        if (getFile().isDown()) {
-            return new DefaultHttpDataSourceFactory(MainApp.getUserAgent(), bandwidthMeter);
-        } else {
-
-            OwnCloudAccount ocAccount = null;
-            try {
-                ocAccount = new OwnCloudAccount(getAccount(), this);
-
-                //Get account credentials asynchronously
-                final PreviewVideoFragment.GetCredentialsTask task = new PreviewVideoFragment.GetCredentialsTask();
-                task.execute(ocAccount);
-
-                OwnCloudCredentials credentials = task.get();
-
-                String login = credentials.getUsername();
-                String password = credentials.getAuthToken();
-
-                Map<String, String> params = new HashMap<String, String>(1);
-
-                if (credentials instanceof OwnCloudBasicCredentials) {
-                    // Basic auth
-                    String cred = login + ":" + password;
-                    String auth = "Basic " + Base64.encodeToString(cred.getBytes(), Base64.URL_SAFE);
-                    params.put("Authorization", auth);
-                } else if (credentials instanceof OwnCloudSamlSsoCredentials) {
-                    // SAML SSO
-                    params.put("Cookie", password);
-                }
-
-                return new CustomHttpDataSourceFactory(MainApp.getUserAgent(), bandwidthMeter, params);
-
-            } catch (AccountUtils.AccountNotFoundException e) {
-                e.printStackTrace();
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            } catch (ExecutionException e) {
-                e.printStackTrace();
-            }
-        }
-
-        return null;
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        if (Util.SDK_INT > 23) {
-            preparePlayer();
-        }
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        if ((Util.SDK_INT <= 23 || player == null)) {
-            preparePlayer();
-        }
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-        if (Util.SDK_INT <= 23) {
-            releasePlayer();
-        }
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-        if (Util.SDK_INT > 23) {
-            releasePlayer();
-        }
-    }
-
-    @Override
-    public void onBackPressed() {
-        Log_OC.v(TAG, "onBackPressed");
-        Intent i = new Intent();
-        i.putExtra(EXTRA_AUTOPLAY, player.getPlayWhenReady());
-        i.putExtra(EXTRA_START_POSITION, player.getCurrentPosition());
-        player.release();
-        setResult(RESULT_OK, i);
-        super.onBackPressed();
-    }
-}
